+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-12-08
+
+### Fixed
+
+-   Support for Refined Storage v2.0.0-milestone.4.11.
+-   Recipe transfer not working in the Wireless Crafting Grid.
+
 ## [0.6.0] - 2024-10-12
 
 ### Added
@@ -78,7 +85,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Recipe transfer integration for the Crafting Grid.
 -   Grid / JEI search box synchronization options.
 
-[Unreleased]: https://github.com/refinedmods/refinedstorage-jei-integration/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/refinedmods/refinedstorage-jei-integration/compare/v0.7.0...HEAD
+
+[0.7.0]: https://github.com/refinedmods/refinedstorage-jei-integration/compare/v0.6.0...v0.7.0
 
 [0.6.0]: https://github.com/refinedmods/refinedstorage-jei-integration/compare/v0.5.0...v0.6.0